+++ src/neo4j/_sync/work/workspace.py
@@ -20,19 +20,26 @@ from __future__ import annotations
 
 import asyncio
 import logging
+import typing as t
 
 from ..._async_compat.util import Util
+from ..._auth_management import to_auth_dict
 from ..._conf import WorkspaceConfig
 from ..._meta import (
     deprecation_warn,
     unclosed_resource_warn,
 )
-from ...api import Bookmarks
+from ...api import (
+    _TAuth,
+    Bookmarks,
+)
+from ...auth_management import AuthManager
 from ...exceptions import (
     ServiceUnavailable,
     SessionError,
     SessionExpired,
 )
+from ..home_db_cache import HomeDbCache
 from ..io import (
     AcquireAuth,
     Neo4jPool,
@@ -83,6 +90,23 @@ class Workspace:
 
     def __exit__(self, exc_type, exc_value, traceback):
         self.close()
+
+    def _make_database_callback(
+        self,
+        auth: t.Union[AuthManager, AuthManager, None]
+    ) -> t.Callable[[str], t.Union[None]]:
+        def _database_callback(database) -> None:
+            nonlocal auth
+            db_cache: HomeDbCache = self._pool.home_db_cache
+            if db_cache.enabled:
+                cache_key = db_cache.compute_key(
+                    self._config.impersonated_user,
+                    self._resolve_session_auth(auth)
+                )
+                db_cache.set(cache_key, database)
+            self._set_cached_database(database)
+
+        return _database_callback
 
     def _set_cached_database(self, database):
         self._cached_database = True
@@ -135,10 +159,8 @@ class Workspace:
 
     def _connect(self, access_mode, auth=None, **acquire_kwargs):
         acquisition_timeout = self._config.connection_acquisition_timeout
-        auth = AcquireAuth(
-            auth,
-            force_auth=acquire_kwargs.pop("force_auth", False),
-        )
+        force_auth=acquire_kwargs.pop("force_auth", False)
+        acquire_auth = AcquireAuth(auth, force_auth=force_auth)
 
         if self._connection:
             # TODO: Investigate this
@@ -146,9 +168,25 @@ class Workspace:
             self._connection.send_all()
             self._connection.fetch_all()
             self._disconnect()
+        self._fill_cached_database(acquire_auth)
+        acquire_kwargs_ = {
+            "access_mode": access_mode,
+            "timeout": acquisition_timeout,
+            "database": self._config.database,
+            "bookmarks": self._get_bookmarks(),
+            "auth": acquire_auth,
+            "liveness_check_timeout": None,
+        }
+        acquire_kwargs_.update(acquire_kwargs)
+        self._connection = self._pool.acquire(**acquire_kwargs_)
+        self._connection_access_mode = access_mode
+
+    def _fill_cached_database(self, acquire_auth: AcquireAuth) -> None:
+        auth = acquire_auth.auth
+        acquisition_timeout = self._config.connection_acquisition_timeout
         if not self._cached_database:
             if (self._config.database is not None
-                    or not isinstance(self._pool, Neo4jPool)):
+                or not isinstance(self._pool, Neo4jPool)):
                 self._set_cached_database(self._config.database)
             else:
                 # This is the first time we open a connection to a server in a
@@ -157,26 +195,47 @@ class Workspace:
                 # to try to fetch the home database. If provided by the server,
                 # we shall use this database explicitly for all subsequent
                 # actions within this session.
+                # Unless we have the resolved home db in out cache:
+
+                db_cache: HomeDbCache = self._pool.home_db_cache
+                cached_db = None
+                if db_cache.enabled:
+                    cache_key = db_cache.compute_key(
+                        self._config.impersonated_user,
+                        self._resolve_session_auth(auth)
+                    )
+                    cached_db = db_cache.get(cache_key)
+                if cached_db is not None:
+                    log.debug("[#0000]  _: <WORKSPACE> resolved home database "
+                              f"from cache: {cached_db}")
+                    self._set_cached_database(cached_db)
+                    return
                 log.debug("[#0000]  _: <WORKSPACE> resolve home database")
                 self._pool.update_routing_table(
                     database=self._config.database,
                     imp_user=self._config.impersonated_user,
                     bookmarks=self._get_bookmarks(),
-                    auth=auth,
+                    auth=acquire_auth,
                     acquisition_timeout=acquisition_timeout,
-                    database_callback=self._set_cached_database
+                    database_callback=self._make_database_callback(auth),
                 )
-        acquire_kwargs_ = {
-            "access_mode": access_mode,
-            "timeout": acquisition_timeout,
-            "database": self._config.database,
-            "bookmarks": self._get_bookmarks(),
-            "auth": auth,
-            "liveness_check_timeout": None,
-        }
-        acquire_kwargs_.update(acquire_kwargs)
-        self._connection = self._pool.acquire(**acquire_kwargs_)
-        self._connection_access_mode = access_mode
+
+    @staticmethod
+    def _resolve_session_auth(
+        auth: t.Union[AuthManager, AuthManager, None]
+    ) -> t.Optional[dict]:
+        if auth is None:
+            return None
+        # resolved_auth = await AsyncUtil.callback(auth.get_auth)
+        # The above line breaks mypy
+        # https://github.com/python/mypy/issues/15295
+        auth_getter: t.Callable[[], t.Union[_TAuth, t.Union[_TAuth]]] = \
+            auth.get_auth
+        # so we enforce the right type here
+        # (explicit type annotation above added as it's a necessary assumption
+        #  for this cast to be correct)
+        resolved_auth = t.cast(_TAuth, Util.callback(auth_getter))
+        return to_auth_dict(resolved_auth)
 
     def _disconnect(self, sync=False):
         if self._connection: